+++ src/Runtime/Optimized/InstanceType/release.ts
@@ -1,0 +1,7 @@
+import type { MmdWasmInstanceType } from "../mmdWasmInstance";
+
+export class MmdWasmReleaseInstanceType implements MmdWasmInstanceType {
+    public getWasmInstanceUrl(): URL {
+        return new URL("../wasm/index_bg.wasm", import.meta.url);
+    }
+}